requirements.txt
steam==1.2.0
motor==2.3.1
aiorcon==0.6.8
discord.py==1.6.0
parsedatetime==2.6
humanize==3.2.0
bidict==0.21.2
aiosqlite==0.17.0
python-dotenv==0.15.0
aiofiles~=0.6.0